Updating compileSdk to 36 and provideValues and com.google.testparameterinjector:test-parameter-injector explicitly to 1.18.
provideValues without providing a context has become deprecated in com.google.testparameterinjector:test-parameter-injector which is used when building your Android app with compileSdk 36.
This is adding a @Suppress("DEPRECATION") annotation the generated code so that no warning is emitted and the code can still be compiled with `allWarningsAsErrors` enabled.

In a future PR the use of provideValues has to be updated.